Move TestEnvironmentOfChildProcess to InterProcessCommunication test Also output redirection for child process replaced with pipes because Console encoding on windows not properly handle non ASCII chars.
